adding two number
In this code, we first declare two integer variables num1 and num2 with values of 5 and 7 respectively. We then add these two numbers together and store the result in a variable called sum. Finally, we print out the result using the System.out.println statement. When we run this code, the output will be:
public class AddTwoNumbers {
   public static void main(String[] args) {
      
      int num1 = 5;
      int num2 = 7;
      int sum = num1 + num2;

      System.out.println("The sum of " + num1 + " and " + num2 + " is " + sum);
   }
}